Work around circular import
